Large refactor
Proc.Send was only being called from tests, so it was replaced with
the logic from SendWithID and the SendWithID method was removed.

On the packeter, the same thing happened with Make - all instances
were specifying an ID, so Make was replaced with the logic from
MakeWithID. But this doesn't need to be exposed and was renamed
to make.

SendQuery and SendResponse were legacy from when messages could be
anything. Now all messages need to be a message.Header so these
were removed and their logic folded into methods on Base.

Removed Chan method from packeter. The channel doesn't need to be
exposed.

The randID function was no longer being used anywhere so it was
removed.